allow affinity mask wider than 32 processors (#3242)
* allow affinity mask wider than 32 processors

the affinity cli option was parsed as int so a mask with any bit above 32
could not be passed at all even though the job stores it as IntPtr and
FixAffinity already handles a 64 bit mask

parsing it as long also means the perfonar model must hold long or the
value gets truncated on the way out

* do not run the wide mask test on a 32 bit runtime

new IntPtr(1L << 40) throws OverflowException where IntPtr is four bytes, so
the test i added would fail on a 32 bit run rather than prove anything. added
a Platform64BitOnly requirement so it skips there, using the same
FactEnvSpecific mechanism the other platform bound tests already use

also added a test for the top bit. the 64th cpu is the last one FixAffinity
handles without cpu groups and its mask only fits in a signed long as the
negative value, so this pins down that it still reaches the right IntPtr

* make the affinity option unsigned

a mask is not a signed number and the 64th cpu needed to be written as
-9223372036854775808 to reach it. now it is written the way the mask reads

the perfonar model stays long. perfolizers LightJsonSerializer throws
Unsupported type: System.UInt64 so making that half unsigned breaks
PerfonarTableTest at runtime. the value is only carried there so a signed long
holds the same bits and nothing is lost

conversion to IntPtr goes through new IntPtr(unchecked((long)value)) which is
the same shape FixAffinity already uses

* reject an affinity mask a 32 bit process can not hold

the option is now unsigned so a value above int.MaxValue can reach
new IntPtr(long) which throws on a 32 bit process instead of
truncating. that made ConfigParser.Parse fail with an unhandled
OverflowException rather than a normal option error.

the conversion now goes through TryConvertAffinity which takes the
pointer size. on 8 bytes it keeps the full 64 bit mask as before. on
4 bytes it takes the low 32 bits so a full 32 processor mask still
works and it reports failure for anything wider so Validate can print
an error.

+++ src/BenchmarkDotNet/Jobs/EnvironmentMode.cs
@@ -148,7 +148,7 @@ namespace BenchmarkDotNet.Jobs
         {
             Jit = HasValue(JitCharacteristic) ? Jit : null,
             Runtime = HasValue(RuntimeCharacteristic) ? Runtime?.RuntimeMoniker : null,
-            Affinity = HasValue(AffinityCharacteristic) ? (int)Affinity : null
+            Affinity = HasValue(AffinityCharacteristic) ? (long)Affinity : null
         };
     }
 }
